Massive reorganization of code and header files
