+++ .gitignore
@@ -1,66 +1,4 @@
-*.libs
-Makefile
-Makefile.in
-aclocal.m4
-autom4te.cache
-config.h
-config.guess
-config.h.in
-config.log
-config.status
-config.sub
-install-sh
-libtool
-libsqltoast-1.0.la
-stamp-h1
-missing
-ltmain.sh
-configure
-depcomp
-sqltoast-1.0.pc
-sqltoast-1.0.pc
-Makefile
-Makefile.in
-aclocal.m4
-autom4te.cache
-config.h
-config.guess
-config.h.in
-config.log
-config.status
-config.sub
-install-sh
-libtool
-libsqltoast-1.0.la
-stamp-h1
-missing
-ltmain.sh
-configure
-*.o
-*.Plo
-*.Po
-*.po
-*.lo
-libsqltoast-1.0.la
-*.deps
-*.dirstamp
-libsqltoast.la
-bin/sqltoast
-config/compile
-config/pandora-plugin.ac
-config/pandora-plugin.am
-config/pandora-plugin.list
-config/pandora_vc_revinfo
-config/plugin.stamp
-docs/header.html
-tests/runtests
-tests/test_value_nodes
-tests/test_manual_ast_construction
-tests/test_lex_create_schema
-bin/example
-tests/test_sql_parser
-tests/test_literals
-tests/test_column_definitions
-tests/table_constraints
-tests/identifiers
-tests/column_definitions
+_build
+_install
+*.~
+*.swp
