--- include/common.h
@@ -1,14 +1,0 @@
-/*
- * Use and distribution licensed under the Apache license version 2.
- *
- * See the COPYING file in the root project directory for full text.
- */
-
-#ifndef SQLTOAST_COMMON_H
-#define SQLTOAST_COMMON_H
-
-namespace sqltoast {
-
-} // namespace sqltoast
-
-#endif /* SQLTOAST_COMMON_H */
