+++ libsqltoast/include/internal/compare.h
