+++ libsqltoast/include/sqltoast.h.in
@@ -1,0 +1,3 @@
+// the configured options and settings for sqltoast
+#define SQLTOAST_VERSION_MAJOR @SQLTOAST_VERSION_MAJOR@
+#define SQLTOAST_VERSION_MINOR @SQLTOAST_VERSION_MINOR@
